Added initial docs for /works response model
